finish renderers

+++ src/main/java/twilightforest/client/state/QuestingRamRenderState.java
@@ -1,12 +1,11 @@
 package twilightforest.client.state;
 
-import net.minecraft.client.renderer.entity.state.EntityRenderState;
 import net.minecraft.client.renderer.entity.state.LivingEntityRenderState;
 import net.minecraft.world.item.DyeColor;
 
 public class QuestingRamRenderState extends LivingEntityRenderState {
 
-	public byte colorFlags;
+	public int colorFlags;
 
 	public boolean isColorPresent(DyeColor color) {
 		return (this.colorFlags & (1 << color.getId())) > 0;
